chore: reorganize the internals folder
