+++ packages/api-rest/internals/package.json
@@ -1,0 +1,8 @@
+{
+	"name": "@aws-amplify/api-rest/internals",
+	"types": "../lib-esm/internals/index.d.ts",
+	"main": "../lib/internals/index.js",
+	"module": "../lib-esm/internals/index.js",
+	"react-native": "../lib-esm/internals/index.js",
+	"sideEffects": false
+}
