--- packages/api-rest/internals/post/package.json
@@ -1,8 +1,0 @@
-{
-	"name": "@aws-amplify/api-rest/internals/post",
-	"types": "../../lib-esm/internalPost.d.ts",
-	"main": "../../lib/internalPost.js",
-	"module": "../../lib-esm/internalPost.js",
-	"react-native": "../../lib-esm/internalPost.js",
-	"sideEffects": false
-}
