+++ packages/api-rest/src/common/handler.ts
